Added ParametricBayesianNetworks
The new class allows specifying a variable number of BayesianNetworks
that are all similar to a default one. The ParametricBayesianNetworks
thus simply stores the default, plus the differences between the default
one and the others. It can then generate them on the fly by simply
constructing a BayesianNetworks with references to the correct values
owned by the ParametricBayesianNetworks.